RFC for converting Session into a protocol and making SessionDriver publicly conformable people can make their own drivers and session types.
This has a number of implications. I'll do my best to provide the reasoning behind these.

Firstly, conforming to a Middleware now means conforming to an instance method instead of a class method. This stemmed originally from trying to restrict SessionDriver to a generic type, but as I pulled that out into just a protocol, I found we would still have to do some odd Swift things to make it work correctly. One thought was to make the SessionDriver a class variable, but it would have to be an optional type since the SessionDriver isn't know until an Application is configured. However, having a variable with shared state like that is not preferable, especially for testability, and I felt it left too much responsibility to the Application class, as it would need to ensure it set that variable instead of simply adding a new Middleware. All in all, I think instance-based middlewares are the way to go, and Providers should be converted to a typealias of a function signature, which would allow closures, global functions, and methods to be used as Providers, so I feel like a divergence between these two concepts is expected.

Secondly, SessionMiddleware is now responsibile for creating the identifier to represent the session. I thought this was appropriate as it is what reads and sets this value into the users cookies, and how we build the session key doesn't really matter to implementors of Session or SessionDriver.

Thirdly, Request's Session reference is now an optional. The original implementation was automatically set to an instance of Session, which was a class, and forced memory-based sessions. Since Sessions are not created until the pass of the SessionMiddleware we can't define the session as non-optional. If we moved session creation outside of the middleware and right before the Request is created we could make it non-optional, but I felt like that was too much for this PR and may move too much responsibility to the ServerDriver, but that can be discussed more.

Looking forward to hearing the feedback on this. I think the next step after making SessionDriver/Session a protocol is figuring out a good way to store other types besides a String inside the session.