Adds the ability to define Middleware that should be ran on a specific controller
Open to comments on this one. This particular feature is inspired by Rails before_filter and Laravel's Controller Middleware. It opens the doors for certain controllers to be behind a logged in user, while others open to everyone. This is a very common need in most web applications, and I'm sure there are plenty more usecases where you'd want a middleware to run on only a certain controller.