Resolves an infinite loop when receiving data
Per the system recv API if the other side of the connection returns 0 bytes the peer has performed an orderly shutdown. Because of this, if Hummingbird's receive returns an empty array, that indicates the connection is closed properly. This updates Vapor code to respect that condition and not continue to loop if the peer is closed.

This issue is causing servers to run at full load. Special thanks to linode for alerting me haha.