Adopt new Hummingbird sockets
I wanted to structure the types in Hummingbird a bit more to allow more robust conformances to C7. Now we have a basic Socket type that handles sending and receiving data, and a ServerSocket that provides the functionality to bind, listen, and accept sockets. There's also a ClientSocket that can perform connections, but that doesn't affect Vapor at all.