Renamed makeJson() to makeJSON()
Initially I created an issue in fluent repository, where I experienced the model  didn't conform to JSONRepresentable.

I figured out that it was due to a typo either in Fluent's Model or in the JSONRepresentable protocol. After awaiting on a response from the creator of the repository I was told it should be makeJSON, which means the typo was in JSONRepresentable.

This PR should fix that typo.