Detect file, function and line of a log call
Currently just the log level and the message are used for logging based
on the given protocol. Where the logging call is actually happening is
ignored which makes debugging much harder.

To fix that, I changed the protocol and all files that consume it in
the main Vapor project to get the file, function and line of a log call
by using “magic” Swift variables. These values can then be used for a
more sophisticated logging and can be ignored if not used.

The position of these variables is important to get the correct
position of the log call in the source code.