[IDEA] Extend CacheProtocol & CacheSessions with expiration support
This commit is not meant to be merged anywhere. It's just a demo of
how CacheProtocol could be extended with an expiration time (seconds),
which could then be used for automatically expiring sessions.

The change in CacheProtocol is breaking as implementations have a new
`set` variant to comply to.
For the demo:
- I added manual expiration to MemoryCache with an inelegant test;
- ignored FluentCache (adding a new column is definitely would be an
issue if there are apps already using this);
- and I'm suggesting that redis-provider (another repo) could use the
SETEX command if an expiration time is provided instead of just SET.