Add refreshCookieEveryRequest to AuthMiddleware
If true, this allows the cookie factory to run every time so that with each request the expiration date can be reset. So, a short expiration date can be set like 10 minutes, but it will be 10 minutes from each successful auth request instead of just 10 minutes from the first login.