Better convert status code into HTTPResponseStatus
This depends on version 1.3.0 of SwiftNIO, see: https://github.com/apple/swift-nio/pull/158

The benefit of this change is that when you have a known status code (e.g. 200) this will be converted to `HTTPResponseStatus.ok` instead of `HTTPResponseStatus.custom(200, "")`. This allows comparing the status code of the resulting `HTTPResponse` directly to a `HTTPResponseStatus` case.