Change drain method's handler parameter closure signature
Moved the `Future` return value as a `Promise` parameter of the handler closure. In addition to updating the route examples, draw your attention to the last example that opens a file handle. Enabled error propagation (`try` instead of `try!`) when closing the file handle (lines 197 and 200) calling it before completing the top-level promise (line 179). See this SO answer https://stackoverflow.com/a/8176030 as to why this is important.