Rework 'SessionData' to be more friendly to 2nd party devs
Motivation:

As it stands right now, `SessionData` is a public type without much use to developers to
write middleware, algorithms, etc. on top of due to how strongly the storage is encapsulated.

It forces all developers to interact with it through `Codable`, as well as not being
testable by allowing developers to validate details stored within it.

Modifications:

- Add: `Equatable` and `ExpressibleByDictionaryLiteral` conformance to `SessionData`
- Add: `SessionData.snapshot` computed property get a copy of the storage
- Change: `SessionData.init(_:)` is now split into two: `init()` and `init(initialData:)`
- Change: `SessionData.init(_:)` is now deprecated

Result:

Developers can now work with `SessionData` in more contexts outside of the Vapor module

This resolves issue #2480