Don't perform double-checked locking in the `Application.http.client.shared` accessor (#2576)
The storage is a Dictionary and can potentially be mutated by one thread while being accessed from another which isn't holding the lock. This is not safe, even if the logic checks again after taking the lock. Unfortunately, this is likely to be a relatively obvious performance regression, but it's also a definitive thread safety concern.