feature: Adds FileMiddleware redirect option
Previously, FileMiddleware would just serve the default file when requesting a directory with no trailing slash. This caused rendering issues when additional resources needed to be loaded by that default file. This adds an option to either redirect to a url with a trailing slash, or just pass the request through the middleware.