Support compiling against Musl
Vapor already makes some provision for compiling against Musl in the
RFC1123 implementation, where `Glibc` is not assumed and is imported
conditionally alongside a conditional import of `Musl`. However, there
are a couple of other places where `Glibc` is still assumed when
compiling for Linux.

This patch replaces these imports with the same `#if canImport(...)`
pattern.