Expose peer certificates in request handlers (#3362)
* Expose peer certificates in request handlers

Additional certificate information can be relevant in mTLS deployments.
Exposing the validated certificate chain to the peer makes them
available per request. This PR exposes the validated peer certificate
chain, i.e., the certificates that establish trust of the peer identity
(from the leaf to and including the root certificate).

* Address API breakage

* Add Swift ASN1 to server tests

---------

Co-authored-by: Tim Condon <0xTim@users.noreply.github.com>